requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx[http2]>=0.28.0
beautifulsoup4>=4.13.0
nltk>=3.9.0
scikit-learn>=1.6.0
rank-bm25>=0.2.2
sentence-transformers>=4.1.0
faiss-cpu>=1.9.0
jinja2>=3.1.0
pydantic>=2.10.0
